+++ index.js
@@ -19,6 +19,7 @@ const camundaCloud10Rules = withConfig({
   'no-multiple-none-start-events': 'error',
   'no-priority-definition': 'error',
   'no-propagate-all-parent-variables': 'error',
+  'no-task-listeners': 'error',
   'no-task-schedule': 'error',
   'no-template': 'error',
   'no-version-tag': 'error',
@@ -94,8 +95,9 @@ const camundaCloud86Rules = withConfig({
 }, { version: '8.6' });
 
 const camundaCloud87Rules = withConfig({
-  ...camundaCloud86Rules,
-  'zeebe-user-task': 'error'
+  ...omit(camundaCloud86Rules, [ 'no-task-listeners' ]),
+  'zeebe-user-task': 'error',
+  'task-listener': 'error'
 }, { version: '8.7' });
 
 const camundaPlatform719Rules = withConfig({
@@ -155,6 +157,7 @@ const rules = {
   'no-propagate-all-parent-variables': './rules/camunda-cloud/no-propagate-all-parent-variables',
   'no-signal-event-sub-process': './rules/camunda-cloud/no-signal-event-sub-process',
   'no-task-schedule': './rules/camunda-cloud/no-task-schedule',
+  'no-task-listeners': './rules/camunda-cloud/no-task-listeners',
   'no-template': './rules/camunda-cloud/no-template',
   'no-version-tag': './rules/camunda-cloud/no-version-tag',
   'no-zeebe-properties': './rules/camunda-cloud/no-zeebe-properties',
@@ -165,6 +168,7 @@ const rules = {
   'signal-reference': './rules/camunda-cloud/signal-reference',
   'start-event-form': './rules/camunda-cloud/start-event-form',
   'subscription': './rules/camunda-cloud/subscription',
+  'task-listener': './rules/camunda-cloud/task-listener',
   'task-schedule': './rules/camunda-cloud/task-schedule',
   'timer': './rules/camunda-cloud/timer',
   'user-task-definition': './rules/camunda-cloud/user-task-definition',